+++ man-roxygen/stats.R
@@ -1,3 +1,3 @@
-#' @param stats.weights Either an integer vector or a list of integer vectors (see [exact_loglik]).
-#' @param stats.statmat Either a matrix or a list of matrices (see [exact_loglik]).
+#' @param stats_weights Either an integer vector or a list of integer vectors (see [exact_loglik]).
+#' @param stats_statmat Either a matrix or a list of matrices (see [exact_loglik]).
 NULL